+++ tests/Commands/InitCommandTest.php
@@ -89,6 +89,7 @@ sonar*
 rector*
 package*
 pint.json
+renovate.json
 *debugbar.json
 ecs*
 *.dist.*